add actions to files
