+++ packages/console/src/services/channel/ConsoleChannel.js
@@ -1,7 +1,7 @@
 import shortid from 'shortid'
 import ClientFileStore from '../storage/ClientFileStore'
 import LocalFileStore from '../storage/LocalFileStore'
-import { splitPath } from 'vtv'
+import { splitPath, joinPath } from 'vtv'
 
 class ConsoleChannel {
   constructor({ name, files }) {
@@ -15,35 +15,7 @@ class ConsoleChannel {
     this.messageIds = []
   }
 
-  async runFilesCommand({ path, args }) {
-    if (path.length === 0) {
-      return {
-        type: 'text',
-        text: 'List not yet implemented',
-      }
-    } else {
-      let result
-      try {
-        result = await this.files.get({
-          path: '/' + path.map(s => encodeURIComponent(s)).join('/'),
-        })
-      } catch (err) {
-        return {
-          type: 'text',
-          text: 'Error getting file',
-        }
-      }
-      return {
-        type: 'tree',
-        name: path[path.length - 1],
-        value: result.body,
-        state: {},
-        message: 'files',
-      }
-    }
-  }
-
-  async runCommand({ message, parsed, onMessage }) {
+  async runCommand({ message, parsed, onMessage, formData }) {
     // TODO: remove once this handles data passed as first parameter
     if (!/^\s*\w/.test(parsed[0].substr(0, 10))) {
       return
@@ -51,24 +23,35 @@ class ConsoleChannel {
 
     const resourcePath = splitPath(parsed[0])
     if (resourcePath[0] === 'files') {
+      let action = null
+      if (formData && formData.action === 'runAction') {
+        action = formData.actionName
+      }
+
       const messageId = shortid()
-      onMessage({
-        type: 'input',
-        text: message,
-        commandId: messageId,
-        loading: true,
-      })
-      const result = await this.runFilesCommand({
+      if (!action) {
+        onMessage({
+          type: 'input',
+          text: message,
+          commandId: messageId,
+          loading: true,
+        })
+      }
+      const result = await this.files.run({
+        action: 'get',
         path: resourcePath.slice(1),
         args: parsed.slice(1),
       })
-      onMessage([
-        {
-          ...result,
-          commandId: messageId,
-        },
-        { type: 'loaded', commandId: messageId },
-      ])
+      onMessage(
+        [
+          {
+            ...result,
+            commandId: messageId,
+            message: joinPath(resourcePath),
+          },
+          { type: 'loaded', commandId: messageId },
+        ].filter(value => value)
+      )
       return result
     }
   }
