+++ packages/console/src/services/storage/actions/index.js
@@ -1,0 +1,7 @@
+import get from './get'
+
+const actions = {
+  get,
+}
+
+export default actions
